server, metrics: move and refine server metrics
